visu (#135)
* fix visu KeyError upon Experiment change

the behaviour stays consisten with what was before 8512bcd. This only
didn't appear because we were looping over the available visualizations,
checking for equality to the 'selected' visualization by hand, which
never happens when the name is empty due to no visualization being
selected. Using this empty name as a key into the dictionary of
available visualizations however does not work.

* simplify visualization configuration