Eliminate redundant data storages and improve moving window view
To eliminate the duplicated storage of all data in each plot dock, the creation and management of the plot docks have been refactored such that the gui holds all the data and only updates the plots.
Furthermore, the plot updates have been changed in such a way that only the internal state of each curve is updated in contrast to all curves being recreated.
This brings the implementation roughly to the state of pymoskito.

Next, the custom moving window code has been removed in favour of pyqtgraphs ViewBox build in xrange functionality.

Last, the pen width is changed to one. This value is recommended in pyqtgraphs settings to improve performance.

Save for the data export which is still to be moved into an extra thread, plotting should be fine now.